fix The given SOAPAction checkVat does not match an operation.
For some unknown reason this gem now throws this error: `The given SOAPAction checkVat does not match an operation.`

The reason for this is that savon inserts a SOAPaction header that brokes it. This workaround removes that header and the error is no longer risen. I didn't find any other workaround to remove this error. For more info look here:

Savon::SOAPRequest#configure_headers